budget_tracker_backend/requirements.txt
wheel==0.38.4
setuptools==67.2.0
fastapi==0.80.0
uvicorn==0.18.2
strawberry-graphql==0.126.2
pytest==7.2.1
pytest-asyncio==0.20.3
asyncio==3.4.3
